requirements.txt
langchain
streamlit
huggingface-hub
sentence_transformers
pinecone-client
python-dotenv
google-generativeai
faiss-cpu
pypdf